feat: migrate to TS
closes #15
closes #28
