--- .flowconfig
@@ -1,8 +1,0 @@
-[include]
-./src/
-
-[ignore]
-.*/node_modules/config-chain/test/
-
-[lints]
-deprecated-utility=off
