+++ .gitignore
@@ -11,21 +11,29 @@ pids
 *.seed
 *.pid.lock
 
+# Diagnostic reports (https://nodejs.org/api/report.html)
+report.[0-9]*.[0-9]*.[0-9]*.[0-9]*.json
+
 # Coverage
 flow-coverage
 lib-cov
 coverage
 .nyc_output
+*.lcov
 
 # Bundle
-build/
-dist/
-target/
+build
+dist
+lib
+target
+
+# Docs
+docs
+doc
 
 # Deps
 node_modules/
 jspm_packages/
-typings/
 flow-typed/npm/
 
 .npm
@@ -33,4 +41,23 @@ flow-typed/npm/
 .node_repl_history
 .env
 
+# IDE
 .idea
+*.iml
+
+# Libdefs
+flow-typed
+typings
+
+# Temp
+temp
+*.tmp
+
+# Typescript
+*.tsbuildinfo
+.tsbuildinfo
+buildcache
+.buildcache
+.rts2_cache_cjs
+.rts2_cache_es
+.rts2_cache_umd
