+++ LICENSE_EN
@@ -1,0 +1,51 @@
+SOFTWARE LICENCE AGREEMENT
+
+Licensor:
+QIWI JSC
+(TIN 7707510721)
+
+2020
+
+According to Art. 1286.1 of the Civil Code of the Russian Federation,
+the Licensor provides an open license for the software product “@qiwi/mixin”
+and related associated documentation files (hereinafter referred to as the “Software”).
+
+Permission is hereby granted, free of charge, to any person obtaining
+a copy of the Software (the "Licensees") to deal in the Software without restriction,
+including without limitation the rights to use, copy, modify, merge, publish, distribute,
+sublicense, and/or sell copies of the Software, and to permit persons to whom
+the Software is furnished in order of distribute/sublicense/sell
+copies of the Software to do so, subject to the following conditions:
+
+The above information about the Licensor and and the provisions of this Agreement
+shall be included in all copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+SOFTWARE.
+
+Except as contained in this Agreement, the name of the Licensor shall not be used
+in advertising or otherwise to promote the sale, use or other dealings
+in the Software without prior written authorization from the Licensor.
+
+This Agreement is a contract of accession and is considered concluded at the time
+of full and unconditional acceptance by the Licensee of its terms. Acceptance
+will be deemed to be the performance by the Licensee of such actions
+that are aimed at fulfilling the conditions (rights and obligations) included in
+the Agreement, including pressing the “Clone or Download” button located in the
+interfaces of the relevant sections on the website https://github.com/qiwi/
+in the Internet telecommunication network, recording/saving/downloading
+the Software (file containing it) into the memory of a computer or other
+technical device (including a mobile phone, smartphone, tablet computer),
+use, copying and/or maintaining the code (part of the code) of the Software
+in a computer memory or other technical devices (including mobile phone, smartphone,
+tablet) in any possible way.
+
+Agreement text is set-up in English and Russian languages.
+Both texts have equal legal force.
+In the case of discrepancy between the texts in Russian and English,
+the Russian text shall prevail.
