+++ LICENSE_RU
@@ -1,0 +1,51 @@
+ЛИЦЕНЗИОННОЕ СОГЛАШЕНИЕ
+
+Лицензиар:
+Акционерное общество «КИВИ»
+(ИНН 7707510721)
+
+2020
+
+Настоящим, согласно ст. 1286.1 Гражданского кодекс Российской Федерации,
+Лицензиар предоставляет открытую лицензию на программный продукт «@qiwi/mixin»
+и сопутствующую документацию (в дальнейшем именуемые «Программное обеспечение»).
+
+Лица, получившие копию Программного обеспечения (далее – «Лицензиаты»), вправе
+безвозмездно использовать Программное обеспечение без ограничений, включая неограниченное
+право на использование, копирование, изменение, слияние, публикацию, распространение,
+сублицензирование и/или продажу копий Программного обеспечения при условии соблюдения ими,
+а также лицами, которым предоставляется данное Программное обеспечение в порядке
+распространения/сублицензирования/продажи копий Программного обеспечения, следующих условий:
+
+Указанные выше данные о Лицензиаре и положения настоящего Соглашения должны быть
+включены во все копии или значимые части данного Программного обеспечения.
+
+ДАННОЕ ПРОГРАММНОЕ ОБЕСПЕЧЕНИЕ ПРЕДОСТАВЛЯЕТСЯ «КАК ЕСТЬ», БЕЗ КАКИХ-ЛИБО ГАРАНТИЙ,
+ЯВНО ВЫРАЖЕННЫХ ИЛИ ПОДРАЗУМЕВАЕМЫХ, ВКЛЮЧАЯ ГАРАНТИИ ТОВАРНОЙ ПРИГОДНОСТИ,
+СООТВЕТСТВИЯ ПО ЕГО КОНКРЕТНОМУ НАЗНАЧЕНИЮ И ОТСУТСТВИЯ НАРУШЕНИЙ, НО НЕ
+ОГРАНИЧИВАЯСЬ ИМИ. НИ В КАКОМ СЛУЧАЕ АВТОРЫ ИЛИ ПРАВООБЛАДАТЕЛИ НЕ НЕСУТ
+ОТВЕТСТВЕННОСТИ ПО КАКИМ-ЛИБО ИСКАМ, ЗА УЩЕРБ ИЛИ ПО ИНЫМ ТРЕБОВАНИЯМ,
+В ТОМ ЧИСЛЕ, ПРИ ДЕЙСТВИИ КОНТРАКТА, ДЕЛИКТЕ ИЛИ ИНОЙ СИТУАЦИИ, ВОЗНИКШИМ ИЗ-ЗА
+ИСПОЛЬЗОВАНИЯ ПРОГРАММНОГО ОБЕСПЕЧЕНИЯ ИЛИ ИНЫХ ДЕЙСТВИЙ С ПРОГРАММНЫМ ОБЕСПЕЧЕНИЕМ.
+
+За исключением случаев, указанных в этом Соглашении, наименование Лицензиара не должно использоваться
+в целях маркетинга или рекламы для расширения продаж, использования или других сделок, связанных
+с этим Программным обеспечением, без предварительного письменного разрешения Лицензиара.
+
+Настоящее Соглашение является договором присоединения и считается заключенным
+в момент полного и безоговорочного принятия Лицензиатом его условий. Акцептом будет
+считаться совершение Лицензиатом таких действий, которые направлены на исполнение
+включенных в Соглашение условий (прав и обязанностей), в том числе, нажатие кнопки
+«Clone or Download», расположенной в интерфейсах соответствующих разделов
+на web-сайте https://github.com/qiwi/ в информационно-телекоммуникационной
+сети «Интернет», запись/сохранение/скачивание Программного обеспечения
+(файла, его содержащего) в память ЭВМ или иного технического устройства
+(включая мобильный телефон, смартфон, планшетный компьютер), использование,
+копирование и/или сохранение кода (части кода) Программного обеспечения в память ЭВМ
+или иного технического устройства (включая мобильный телефон, смартфон,
+планшетный компьютер) любым доступным способом.
+
+Текст настоящего Соглашения составлен на английском и русском языках.
+Оба текста имеют равную юридическую силу.
+В случае расхождения между текстом на русском и английском языках,
+версия на русском языке имеет преимущественную силу.
