+++ package.json
@@ -3,31 +3,39 @@
   "version": "1.2.4",
   "description": "Health indicator kit for server-side monitoring & balancing",
   "main": "target/es5/index.js",
-  "directories": {
-    "source": "src/main",
-    "test": "src/test"
+  "source": "src/main/ts/index.ts",
+  "types": "typings/index.d.ts",
+  "typescript": {
+    "definition": "typings/index.d.ts"
   },
   "files": [
     "README.md",
     "CHANGELOG.md",
     "target",
+    "typings",
     "flow-typed"
   ],
-  "@std/esm": {
-    "mode": "all",
-    "cjs": true
-  },
   "scripts": {
-    "flow": "flow",
-    "flow:coverage": "flow-coverage-report -i 'src/main/**/*.js' -t html -t json -t text --threshold 90",
-    "libdefs:fetch": "flow-typed install lodash-es@4.17.11",
-    "jest": "jest --config=jest.config.json",
-    "lint": "standard-flow src/**/*.js --fix",
-    "test": "BABEL_ENV=test yarn flow:coverage && yarn lint && yarn jest",
-    "build": "rm -rf ./target; BABEL_ENV=production babel src/main/js --out-dir target/es5",
-    "test:report": "yarn test && yarn coveralls:push",
+    "jest": "jest -w 2 --config=jest.config.json",
+    "test": "yarn lint && yarn jest",
+    "clean": "rimraf target typings",
+    "build": "yarn clean && yarn build:es5 && yarn build:es6 && yarn build:ts && yarn libdef && yarn docs && yarn uglify && yarn build:bundle",
+    "build:es5": "mkdir -p target/es5 && tsc -p tsconfig.es5.json",
+    "build:es6": "mkdir -p target/es6 && tsc -p tsconfig.es6.json",
+    "build:ts": "cp -r src/main/ts/ target/ts/",
+    "build:bundle": "microbundle build src/main/ts/index.ts -o target/bundle",
+    "dtsgen": "dts-generator --project ./ --out typings/index.d.ts --prefix @qiwi/health-indicator/target/es5 --name @qiwi/health-indicator --main @qiwi/health-indicator/target/es5/index --moduleResolution node",
+    "flowgen": "flowgen typings/index.d.ts --output-file flow-typed/index.flow.js",
+    "libdef": "yarn dtsgen && yarn libdeffix && yarn flowgen",
+    "libdeffix": "node -r esm ./scripts/js/libdef-fix.js --dts=./typings/index.d.ts",
+    "test:report": "yarn test && yarn push:report",
+    "push:report": "yarn coveralls:push",
     "coveralls:push": "cat ./coverage/lcov.info | coveralls",
-    "postupdate": "yarn && yarn build && yarn test"
+    "docs": "typedoc --readme README.md --tsconfig tsconfig.json src/main --ignoreCompilerErrors || exit 0",
+    "uglify": "for f in $(find target -name '*.js'); do short=${f%.js}; terser -c -m -o $short.js -- $f; done",
+    "postupdate": "yarn && yarn build && yarn test",
+    "lint": "eslint src/**/*.ts",
+    "lint:fix": "yarn lint --fix"
   },
   "repository": {
     "type": "git",
@@ -44,40 +52,47 @@
   "license": "MIT",
   "dependencies": {
     "http-status-codes": "^1.4.0",
-    "lodash-es": "^4.17.15"
+    "lodash": "^4.17.15"
   },
   "devDependencies": {
     "@babel/cli": "^7.8.4",
     "@babel/core": "^7.8.4",
+    "@babel/plugin-external-helpers": "^7.8.3",
     "@babel/plugin-proposal-class-properties": "^7.8.3",
-    "@babel/plugin-proposal-object-rest-spread": "^7.8.3",
-    "@babel/plugin-transform-modules-commonjs": "^7.8.3",
+    "@babel/plugin-transform-runtime": "^7.8.3",
+    "@babel/polyfill": "^7.8.3",
     "@babel/preset-env": "^7.8.4",
-    "@babel/preset-flow": "^7.8.3",
     "@babel/register": "^7.8.3",
+    "@babel/runtime": "^7.8.4",
+    "@qiwi/license": "^1.1.1",
+    "@qiwi/semantic-release-gh-pages-plugin": "^1.15.8",
     "@semantic-release/changelog": "^5.0.0",
     "@semantic-release/git": "^9.0.0",
-    "@semantic-release/github": "^7.0.1",
+    "@semantic-release/github": "^7.0.2",
     "@semantic-release/npm": "^7.0.2",
-    "babel-core": "^7.0.0-bridge.0",
-    "babel-jest": "^25.1.0",
-    "babel-eslint": "^10.0.3",
-    "babel-preset-minify": "^0.5.1",
-    "chai": "^4.2.0",
-    "chai-spies": "^1.0.0",
-    "chai-subset": "^1.6.0",
+    "@types/jest": "^25.1.1",
+    "@types/jest-json-schema": "^2.1.1",
+    "@typescript-eslint/parser": "^2.19.0",
     "coveralls": "^3.0.9",
+    "dts-generator": "^3.0.0",
     "eslint": "^6.8.0",
-    "eslint-config-standard": "^14.1.0",
-    "eslint-config-standard-flow": "^1.0.1",
-    "flow-bin": "^0.117.0",
-    "flow-coverage-report": "^0.6.1",
-    "flow-remove-types": "^2.117.1",
-    "flow-typed": "^2.6.2",
+    "eslint-config-qiwi": "^1.3.1",
+    "eslint-plugin-typescript": "^0.14.0",
+    "esm": "^3.2.25",
+    "flowgen": "1.10.0",
     "jest": "^25.1.0",
-    "reqresnext": "^1.6.2",
+    "lodash": "^4.17.15",
+    "microbundle": "^0.12.0-next.6",
+    "replace-in-file": "^5.0.2",
+    "reqresnext": "^1.6.3",
+    "rimraf": "^3.0.1",
     "semantic-release": "^17.0.2",
-    "standard-flow": "^1.0.0"
+    "terser": "^4.6.3",
+    "ts-jest": "^25.1.0",
+    "typedoc": "0.16.9",
+    "typedoc-plugin-external-module-name": "^3.0.0",
+    "typescript": "^3.7.5",
+    "typescript-eslint-parser": "^22.0.0"
   },
   "release": {
     "branch": "master",
@@ -95,22 +110,5 @@
       "@semantic-release/npm",
       "@semantic-release/github"
     ]
-  },
-  "standard-flow": {
-    "parser": "babel-eslint",
-    "globals": [
-      "it",
-      "expect",
-      "describe"
-    ],
-    "ignore": [],
-    "plugins": [
-      "flowtype"
-    ],
-    "settings": {
-      "flowtype": {
-        "onlyFilesWithFlowAnnotation": true
-      }
-    }
   }
 }
