--- src/main/js/base/index.js
@@ -1,21 +1,0 @@
-// @flow
-
-export {
-  extend,
-  minBy,
-  maxBy,
-  each,
-  find,
-  isEmpty,
-  isUndefined,
-  isObject,
-  isString,
-  isFunction,
-  mapValues,
-  toArray,
-  pickBy
-} from 'lodash-es'
-
-export function isDefined (value?: any): boolean {
-  return value !== undefined
-}
