--- src/main/js/endpoint/statuses.js
@@ -1,8 +1,0 @@
-// @flow
-
-export {
-  OK,
-  NOT_FOUND,
-  SERVICE_UNAVAILABLE,
-  INTERNAL_SERVER_ERROR
-} from 'http-status-codes'
