--- src/main/js/index.js
@@ -1,6 +1,0 @@
-// @flow
-
-import Endpoint from './endpoint'
-
-export * from './indicator'
-export {Endpoint}
