--- src/main/js/indicator/index.js
@@ -1,7 +1,0 @@
-// @flow
-
-import AbstractIndicator from './abstract'
-import StandardIndicator from './standard'
-import SemaphoreIndicator from './semaphore'
-
-export {AbstractIndicator, StandardIndicator, SemaphoreIndicator}
