+++ src/main/ts/endpoint/statuses.ts
@@ -1,0 +1,6 @@
+export {
+  OK,
+  NOT_FOUND,
+  SERVICE_UNAVAILABLE,
+  INTERNAL_SERVER_ERROR
+} from 'http-status-codes'
