+++ src/main/ts/index.ts
@@ -1,0 +1,4 @@
+import Endpoint from './endpoint'
+
+export * from './indicator'
+export { Endpoint }
