+++ src/main/ts/indicator/index.ts
@@ -1,0 +1,5 @@
+import AbstractIndicator from './abstract'
+import StandardIndicator from './standard'
+import SemaphoreIndicator from './semaphore'
+
+export { AbstractIndicator, StandardIndicator, SemaphoreIndicator }
