--- src/test/js/index.js
@@ -1,12 +1,0 @@
-import {AbstractIndicator, SemaphoreIndicator, StandardIndicator, Endpoint} from '../../main/js'
-import chai from 'chai'
-
-const {expect} = chai
-
-describe('index', () => {
-  it('properly exposes its inners', () => {
-    ([AbstractIndicator, SemaphoreIndicator, StandardIndicator, Endpoint]).forEach(item => {
-      expect(item).not.to.be.undefined
-    })
-  })
-})
